remove @radix-ui/react-icons & lazy load navbar components & load scheherazade font only in reader
