+++ src/lib/fonts.ts
@@ -1,12 +1,5 @@
-import { Plus_Jakarta_Sans, Scheherazade_New } from "next/font/google";
+import { Plus_Jakarta_Sans } from "next/font/google";
 import localFont from "next/font/local";
-
-const scheherazade = Scheherazade_New({
-  subsets: ["arabic"],
-  style: "normal",
-  variable: "--font-scheherazade",
-  weight: ["400", "700"],
-});
 
 const plusJakartaSans = Plus_Jakarta_Sans({
   subsets: ["latin"],
@@ -78,9 +71,4 @@ const ibmPlexSansArabic = localFont({
 // });
 
 export const getFontsClassnames = () =>
-  [
-    plusJakartaSans.variable,
-    ibmPlexSansArabic.variable,
-    scheherazade.variable,
-    // uthmanicHafs.variable,
-  ].join(" ");
+  [plusJakartaSans.variable, ibmPlexSansArabic.variable].join(" ");
